test(core): relocate/add tests so core reaches full own-crate coverage
Move the pure value-type interaction tests out of leviath-cli into a new
tests module in leviath-core/src/interaction.rs (the value types moved to
core during the relocation refactor but their tests were left in cli, so
they showed as uncovered under `-p leviath-core`). The cli file-IPC/stdin
tests stay in cli.

Add TitleConfig tests to core/config.rs, plus targeted manifest.rs tests
(non-string tool_routing override, full parse_security_config branch
coverage) and a region.rs test for the evict_lru_entry empty-guard.

interaction.rs and config.rs reach 100%; manifest.rs and region.rs are at
their unreachable-defensive-branch floor (serde_json::to_value never errs
on TOML values; remove_oldest never returns None while a sliding window is
over budget).
